Moved tailwind plugins into own file| Use lodash|  Reformatted files| Changed brand
