+++ frontend/config/design/helpers/hex-to-rgb.js
@@ -1,13 +1,12 @@
 module.exports = function (hex) {
-    var c
-    if (/^#([A-Fa-f0-9]{3}){1,2}$/.test(hex)) {
-        c = hex.substring(1).split('')
-        if (c.length == 3) {
-            c = [c[0], c[0], c[1], c[1], c[2], c[2]]
-        }
-        c = '0x' + c.join('')
-        return [(c >> 16) & 255, (c >> 8) & 255, c & 255, 1]
-    } else {
-        return null
+  if (/^#([A-Fa-f0-9]{3}){1,2}$/.test(hex)) {
+    let c = hex.substring(1).split('')
+    if (c.length === 3) {
+      c = [c[0], c[0], c[1], c[1], c[2], c[2]]
     }
+    c = '0x' + c.join('')
+    return [(c >> 16) & 255, (c >> 8) & 255, c & 255, 1]
+  } else {
+    return null
+  }
 }
