+++ frontend/config/design/plugins/tailwind-plugins.js
@@ -1,0 +1,21 @@
+module.exports = [
+  function({addUtilities, config}) {
+    let controlHeight = parseInt(config('height.control-height').split('p')[0])
+    const newUtilities = {
+      '.text-control-font-size': {
+        fontSize: controlHeight * .3 + 'px',
+      },
+      '.leading-control-line-height': {
+        lineHeight: controlHeight + 'px'
+      },
+      '.px-control-x': {
+        'paddingLeft': controlHeight * .4 + 'px',
+        'paddingRight': controlHeight * .4 + 'px'
+      }
+    }
+    addUtilities(newUtilities, {
+      respectPrefix: false,
+      respectImportant: false,  
+    })
+  }
+]
